feat: changed file structer, removed the threadlifetime and made minor chanage
