+++ mava/utils/sebulba/utils.py
@@ -1,0 +1,58 @@
+
+
+import queue
+import threading
+import time
+import jax
+from mava.systems.ppo.types import Params
+from typing import Any 
+
+class RecordTimeTo:
+    """Context manager to record the runtime in a `with` block"""
+
+    def __init__(self, to: Any):
+        self.to = to
+
+    def __enter__(self) -> None:
+        self.start = time.monotonic()
+
+    def __exit__(self, *args: Any) -> None:
+        end = time.monotonic()
+        self.to.append(end - self.start)
+
+class ParamsSource(threading.Thread):
+    """A `ParamSource` is a component that allows networks params to be passed from a
+    `Learner` component to `Actor` components.
+    """
+
+    def __init__(self, init_value: Params, device: jax.Device):
+        super().__init__(name=f"ParamsSource-{device.id}")
+        self.value: Params = jax.device_put(init_value, device)
+        self.device = device
+        self.new_value: queue.Queue = queue.Queue()
+        self._stop_event = threading.Event()
+
+    def run(self) -> None:
+        """This function is responsible for updating the value of the `ParamSource` when a new value
+        is available.
+        """
+        while not self._stop_event.is_set():
+            try:
+                waiting = self.new_value.get(block=True, timeout=1)
+                self.value = jax.device_put(waiting, self.device)
+            except queue.Empty:
+                continue
+
+    def update(self, new_params: Params) -> None:
+        """Update the value of the `ParamSource` with a new value.
+
+        Args:
+            new_params: The new value to update the `ParamSource` with.
+        """
+        self.new_value.put(new_params)
+
+    def get(self) -> Params:
+        """Get the current value of the `ParamSource`."""
+        return self.value
+
+
